Added intial controls for camera and cursor
Code needs to be cleaned up in GameDriver

Added control classes. They need to be finalized and updated in GameControlMapper

Updated ContentLoader to fetch the game cursor textures

Updated MapContainer and TmxMapParser so that they pass the grid instead of creating a map entirely. Moved map initialization to the GameDriver for now.